+++ NEWS.md
@@ -1,6 +1,7 @@
 # treeio 1.3.14
 
-+ read.iqtree (2018-04-17, Tue)
++ read.astral (2018-04-17, Tue)
++ read.iqtree
 
 # treeio 1.3.13