+++ NEWS.md
@@ -1,5 +1,6 @@
 # treeio 1.3.14
 
++ rename_taxa (2018-04-19, Thu)
 + read.astral (2018-04-17, Tue)
 + read.iqtree